+++ example.py
@@ -29,8 +29,9 @@ for epoch in range(3):
         loss = model.train_on_batch(x, y)
         eta_history.append(K_eval(model.optimizer.eta_t, K))
         print("Iter {} loss: {}".format(iteration + 1, "%.3f" % loss))
+        if iteration == (24 - 2):
+            K.set_value(model.optimizer.t_cur, -1)  # WARM RESTART
     print("EPOCH {} COMPLETED\n".format(epoch + 1))
-    K.set_value(model.optimizer.t_cur, 0)  # WARM RESTART
 
 plt.plot(eta_history, linewidth=2)
 plt.xlim(0, len(eta_history))
