+++ keras_adamw/__init__.py
@@ -29,4 +29,4 @@ from .utils import get_weight_decays, fill_dict_in_order
 from .utils import reset_seeds, K_eval
 
 
-__version__ = '1.25'
+__version__ = '1.3'
