+++ tests/__init__.py
@@ -1,3 +1,1 @@
 from .import_selection import *
-
-__version__ = '1.1'
