--- tests/test_optimizers/__init__.py
@@ -1,1 +1,0 @@
-from .test_optimizers import *
