--- tests/test_optimizers_225/__init__.py
@@ -1,1 +1,0 @@
-from .test_optimizers import *
