--- tests/test_optimizers_225tf/__init__.py
@@ -1,1 +1,0 @@
-from .test_optimizers_v2 import *
